requirement.txt
request
pandas
PyYAML
python-dotenv
mysql-connector-python
beautifulsoup4
pyspark
uszipcode